+++ BranchSDK/src/BranchIO/Util/Sleeper.cpp
@@ -2,7 +2,7 @@
 
 #include <Poco/Clock.h>
 
-#include "BranchIO/Sleeper.h"
+#include "Sleeper.h"
 
 using namespace std;
 using namespace Poco;
